Implement binder for rkpdapp IRegistration AIDL
This contains a nearly complete implementation of the
com.android.rkpdapp.IRegistration interface for getting keys from rkpd.

The one missing piece of the implementation is the background work that
needs to be scheduled when keys are consumed. As the background worker
is not yet committed, this work will be completed under bug 262253838.

Test: RkpdAppUnitTests
Bug: 260742161
Change-Id: I1ed079c33340713e1048e2b07883c18b9b23ec84