Add callback to return status for storeUpgradedKey
Failing open (ignoring errors) on storeUpgradedKey will leave the
device open to potential downgrade attacks. Address this by adding
result handling for storeUpgradedKey in the form of a callback.

This async call is not cancelable by design. It only does some local
file I/O and should not block on any other resources.

Test: RkpdAppUnitTests (new tests added in this CL)
Bug: 262748535
Change-Id: I636a96d64decbdc2f16108df5d61899df1cd2fe7