Fix a race in rkpd getkey API
ProvisionedKeyDao was a bit buggy in that concurrent calls could end
up violating database constraints. This is because the method didn't
handle cases in which two threads try to assign a key to the same
caller at the same time. In this case, both callers would find an
unassigned key then update it in the database. This violates
constraints, causing a crash.

Change assignKey to getOrAssignKey so that it removes the race by always
returning a previously-assigned key before trying to assign a fresh key.

Test: RkpdAppIntegrationTest, RkpdAppStressTest, RkpdAppUnitTest
Bug: 266483337
Change-Id: I8f40a672c29438ab7a17b6f935b0d5305caa334e