Fix test flakes in ProvisionedKeyTest
It turns out our random input data was not random enough to prevent
collisions on hashCode tests. Increase the size of the random input
data to at least 16 bits so the chances of collision should be near-0.

Tested by adding a loop around the previously flaky test and running it
a million times.

Fixes: 266731370
Test: RkpdAppUnitTests
Change-Id: I6b883b58e3e6d26c9043aa8fb8028e3497ee7408