Replace "get expiring keys" with a method to get a count
We never need the list of expiring keys, so it was wasteful to ever
actually read the data out of the db. Instead, we just need to get
a count of the keys that are expiring, which should be a less costly
operation.

Expand the unit test for this method a bit to verify more cases.

Test: RkpdAppIntegrationTests RkpdAppUnitTests RkpdAppStressTests
Change-Id: I371f0c3b962bd78a7a5ae659f7fdbdd28b817d75