Use IBinder instead of Proxy object for callback equivalence
When canceling a callback, the object is passed into our binder
implementation. We were comparing the callbacks directly, but there
is no guarantee that the same remote binder will be comparable via
the callback types. Instead, call asBinder to get the underlying
binder. This is guaranteed to be the same when there are multiple
calls into our backend from the same client.

Bug: 275653768
Test: RkpdAppUnitTests RkpdAppIntegrationTests RkpdAppHostTests
Test: RemoteProvisioningServiceTests keystore2_test
Change-Id: I5d67068ffb65498ba290e7e2d9cda0ceb94be0fc
Merged-In: I5d67068ffb65498ba290e7e2d9cda0ceb94be0fc