Do not swallow InterruptedException in ServerInterface
We rely on thread interruption to properly cancel async operations.
A new test, testCancelDueToServiceTimeout, was flaky because were
swallowing an InterruptedException in the ServerInterface retry loop,
so we might sometimes retry a server call when, in fact, the operation
should have been canceled.

Bug: 275653768
Test: RkpdAppUnitTests RkpdAppIntegrationTests RkpdAppHostTests
Test: RemoteProvisioningServiceTests keystore2_test
Change-Id: Ia374ef3e9b9413b1aa7a3788f4064546545a7021
Merged-In: Ia374ef3e9b9413b1aa7a3788f4064546545a7021