Accept PENDING_INTERNET_CONNECTIVITY as a TRANSIENT_ERROR
Really either of OUT_OF_KEYS_TRANSIENT_ERROR or
OUT_OF_KEYS_PENDING_INTERNET_CONNECTIVITY are acceptable in this test
case. Trying to detect network connectivity to select from either of
these enum values had a race condition:

* Connection goes down
* Test runs
* Connection goes up
* Checks connection status
* Asserts error code based on connection status
X TEST FAIL

These two error codes are there in case a client wants to behave
differently when seeing one instead of the other. This test however is
trying to derive one bit of state from the range of possible error code
values (that bit being "Are we out of keys - but not from a terminal
failure state?").

Bug: 288523573
Test: Presubmit
Change-Id: I704f55ad19c5af77237ec948e5eef389c847832d