Skip periodic provisioner tests on devices with RKP disabled
The periodic provisioner noops if RKP is disabled (no default hostname).
This causes most of the unit tests to fail, as they actually need
the provisioner to do something.

Add an assumption to skip these tests when RKP is disabled.

Ignore-AOSP-First: A slightly more complex fix will be submitted to aosp
Test: adb install RkpdAppGoogleUnitTests.apk && am shell instrument ...
Bug: 289986606
Change-Id: I2ae48d2d4abd71fd649dbe7c977596df6b297a8d