Reduce fragility of PeriodicProvisionerTests
We were using verifyNoMoreInteractions, which ensures that NO calls to a
HAL were made. This is far too strict because there's a valid chance
some methods are called. For example, `getServiceName` might be called
due to some call to `ServiceManagerInterface`, which iterates over the
HALs, but never does meaningful operations with them.

Instead, verify that the most important call, `generateKey`, is never
invoked. This is a more precise way to check that a HAL was not used.

Bug: 294851387
Test: RkpdAppIntegrationTests
Change-Id: Ic6f0a881f802ef75d74fb8465c48c9b3a7260579