Embed Request ID into `GeekResponse`
Add a Request ID to GeekResponse and make it available in the class. The
requestID can be reused for fetchEEK, SignCertificates, and later for
ConfirmCertificates. Generating UUID is cheap, so it's alright to create
it internally than have a setter for it. Later, if
`enable_request_id_reuse` flag is set, we'd use this requestID value
instead of creating a new one in ServerInterface that we currently do.

Flag: EXEMPT in preparation for using the flag enable_request_id_reuse
Bug: 389829172
Change-Id: If6f2ba1d10db9b06dbc41523c895b6b9956770c2