Move NetworkUtils to a shared utility library.
The `NetworkUtils` class, which checks for network consent, is moved
from the RKPD app package to a new `rkp-util` library in the `util`
directory. The `RegistrationProxy` in the system server now uses this
utility to ensure network consent is present before attempting to
connect to the RKPD service, throwing a new error code if consent is not
assumed. This fixes the issue where RKPD application was being launched
even when it cannot perform any operation, causing loss of resources.

Bug: 442528393
Test: atest RkpdAppGoogleUnitTests RkpdAppGoogleIntegrationTests
Flag: EXEMPT BUGFIX
Change-Id: I43dc6a01e8086235c0750a73f3b1ea160b4b1eee